analyst/requirements.txt
strands-agents>=1.0.0
feedparser>=6.0.10
requests>=2.31.0
beautifulsoup4>=4.12.0
readability-lxml>=0.8
pyyaml>=6.0